option to put emails with constant login codes info cfg (#211)

+++ kyc/face/internal/threedivi/threedivi.go
@@ -113,7 +113,7 @@ func (t *threeDivi) Available(_ context.Context, userWasPreviouslyForwardedToFac
 
 //nolint:revive // .
 func (t *threeDivi) isAvailable(userWasPreviouslyForwardedToFaceKYC bool) error {
-	if int64(t.cfg.ThreeDiVi.ConcurrentUsers)-(int64(t.activeUsersCount.Load())+int64(t.loadBalancedUsersCount.Load())) >= 1 {
+	if uint64(t.cfg.ThreeDiVi.ConcurrentUsers)-(t.activeUsersCount.Load()+t.loadBalancedUsersCount.Load()) >= 1 {
 		if !userWasPreviouslyForwardedToFaceKYC {
 			t.loadBalancedUsersCount.Add(1)
 		}
